Widen GenericForeignKey object_id fields to bigint, document schema state
WorkflowInstance.object_id and UserNotification.object_id were 32-bit
PositiveIntegerField while several entities they point to (TravelRequest,
etc.) use bigint primary keys - would silently wrap/fail once a
referenced id exceeds ~2.1 billion. Widened both to
PositiveBigIntegerField; safe, no data migration needed.

Also documents the schema cleanup from the last few commits in
ARCHITECTURE.md: a new section on database conventions (PK typing, the
dropped legacy/expenses tables, the intentional dual-write between the
per-module approval-step tables and the generic workflow engine, the
approver_role/approver_permission transition), plus a note in the RBAC
section clarifying that Django's built-in Group/Permission tables are
unused scaffolding, not a second live permission system.

+++ backend/workflows/models.py
@@ -327,7 +327,12 @@ class WorkflowInstance(models.Model):
 
     # Generic relation to any entity (TRF, Visa, etc.)
     content_type = models.ForeignKey(ContentType, on_delete=models.CASCADE)
-    object_id = models.PositiveIntegerField()
+    # PositiveBigIntegerField, not PositiveIntegerField (ERD fix roadmap,
+    # Issue 14): several entities this can point to (trf_travelrequest,
+    # visa_visaapplication, etc.) use bigint PKs, which a 32-bit
+    # PositiveIntegerField would silently wrap/fail on once one exceeds
+    # ~2.1 billion.
+    object_id = models.PositiveBigIntegerField()
     content_object = GenericForeignKey("content_type", "object_id")
 
     # Workflow state
